Start from scratch, this is the proposed starting point.
