--- MyFirstApp/app/src/androidTest/java/com/example/myfirstapp/UITest2.java
@@ -1,35 +1,0 @@
-package com.example.myfirstapp;
-
-
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-import androidx.test.filters.LargeTest;
-import androidx.test.rule.ActivityTestRule;
-
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-//import static androidx.test.espresso.Espresso.onView;
-import static androidx.test.espresso.Espresso.onView;
-import static androidx.test.espresso.action.ViewActions.click;
-import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
-import static androidx.test.espresso.action.ViewActions.typeText;
-import static androidx.test.espresso.assertion.ViewAssertions.matches;
-import static androidx.test.espresso.matcher.ViewMatchers.withId;
-import static androidx.test.espresso.matcher.ViewMatchers.withText;
-
-@RunWith(AndroidJUnit4.class)
-@LargeTest
-public class UITest2 {
-    @Rule
-    public ActivityTestRule<MainActivity> activityRule =
-            new ActivityTestRule<>(MainActivity.class);
-
-    @Test
-    public void listGoesOverTheFold() {
-        onView(withId(R.id.editText)).perform(typeText("This is a Test"), closeSoftKeyboard());
-        onView(withId(R.id.button)).perform(click());
-        onView(withId(R.id.textView)).check(matches(withText("This is a Test")));
-    }
-
-}
