--- MyFirstApp/app/src/main/java/com/example/myfirstapp/DataStorage/LocalStorage.java
@@ -1,13 +1,0 @@
-package com.example.myfirstapp.DataStorage;
-
-public class LocalStorage {
-    String data;
-
-    public void setData(String data) {
-        this.data = data;
-    }
-
-    public String getData() {
-        return data;
-    }
-}
