--- MyFirstApp/app/src/test/java/com/example/myfirstapp/ExampleUnitTest.java
@@ -1,25 +1,0 @@
-package com.example.myfirstapp;
-
-import com.example.myfirstapp.DataStorage.LocalStorage;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
-    }
-    @Test
-    public void localstorage() {
-        LocalStorage localStorage = new LocalStorage();
-        localStorage.setData("testing");
-        assertEquals("testing", localStorage.getData());
-    }
-}
