--- MyFirstApp/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Wed Sep 09 10:10:33 PDT 2020
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.1.1-all.zip
