+++ app/src/res/mipmap-xxxhdpi/ic_launcher.png
